Remove synthetics `BaseCommand` class
